Added booking tabs
